Removing code that triggers footer pop up
There was a pop up in the footer with the class "small-dialog" in
footer.php. That footer has been removed and this code causes an error
on the screen that says "Content Not Found". Removing this code and the
footer pop up resolve the issue. Footer Pop up was replaced with the
Travel Guide 2019 popup using Convert Plus Plugin